Streamlit-apps2/requirements.txt
selenium
beautiful_soup
tweepy
streamlit
webdriver_manager
pandas
levenshtein
fuzzywuzzy
tweet-preprocessor
spacy
seaborn
matplotlib
vaderSentiment
scikit-learn